fix: remove `github_token` default
We need to get access to Projects V2 through the API, and there's no way
to do that with repo-scoped tokens, since V2 projects live on
users/organizations.
For this reason, it makes no sense to use the actions token as a default
value, since it's impossible for it to work.